ALlow rejecting uids below a certain number
We do not want the github user 'root' or 'www-data' to be
able to ssh into any host. We 'enforce' this by requiring that
all sshing users have a uid of > 999 by default (can be
changed with the --min-uid commandline flag).

This is a first line of defence and should make my current
running server be not terribly insecure. Need more configuration
knobs, and eventually a pure whitelist option.